+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2021-03-27
+## [Unreleased] - 2021-03-29
 
 ### Added
+* New methods added to the Permutation class.
+    * The cycle method creates a permutation cycle.
 
 ### Changed